DESIGN : 인기 공고문 UI 수정
